Move AutomateAuthToken to the common crate
This will allow us to refer to it from both `hab` (for the CLI) and
`hab-sup`, without introducing an unwanted dependency on `hab-sup` for
`hab`
